backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pymongo
motor
python-dotenv
pydantic[email]
bcrypt
python-jose
google
google.generativeai
python-multipart